Add basic Makefile
This Makefile is notably missing an important target: install

Invoking make at a command line builds a binary that can be run
directly from the repository's directory. The binary is wholly
unsuited for running in a different directory, such as the way
Linux installations typically do this.

One way to solve this would be as follows:
Modify all path references vary based on a constant variable in
a header file. This has the advantage of easily being overwritten
by a Make target.

Another way, and less advisable, is to install everything under
/usr/libexec/Game-Pencil-Engine/ including the binary, and provide
a shell script shim to be installed in /usr/bin that simply calls
/usr/libexec/Game-Pencil-Engine/Game_Pencil_Engine_IDE